+++ Core/ModuleInstaller.cs
@@ -1092,7 +1092,8 @@ namespace CKAN
         /// <param name="files">Set of files to import</param>
         /// <param name="user">Object for user interaction</param>
         /// <param name="installMod">Function to call to mark a mod for installation</param>
-        public void ImportFiles(HashSet<FileInfo> files, IUser user, Action<string> installMod)
+        /// <param name="allowDelete">True to ask user whether to delete imported files, false to leave the files as is</param>
+        public void ImportFiles(HashSet<FileInfo> files, IUser user, Action<string> installMod, bool allowDelete = true)
         {
             Registry         registry    = registry_manager.registry;
             HashSet<string>  installable = new HashSet<string>();
@@ -1134,7 +1135,7 @@ namespace CKAN
                 }
                 ++i;
             }
-            if (installable.Count > 0 && user.RaiseYesNoDialog($"Install {installable.Count} compatible imported mods?"))
+            if (installable.Count > 0 && user.RaiseYesNoDialog($"Install {installable.Count} compatible imported mods in game instance {ksp.Name} ({ksp.GameDir()})?"))
             {
                 // Install the imported mods
                 foreach (string identifier in installable)
@@ -1142,7 +1143,7 @@ namespace CKAN
                     installMod(identifier);
                 }
             }
-            if (user.RaiseYesNoDialog($"Import complete. Delete {deletable.Count} old files?"))
+            if (allowDelete && deletable.Count > 0 && user.RaiseYesNoDialog($"Import complete. Delete {deletable.Count} old files?"))
             {
                 // Delete old files
                 foreach (FileInfo f in deletable)